Update jquery.extend.js
Added debug object (persists to local storage) and overridden console functions so they only work when debug is enabled.  This allows you to put debugging information in all your code and enable and disable it on a live site via the console.

Also added debug logging for all ajax requests.